Reworked live patching routine in DIMM_UploadFile function. Definitely more efficient and well thought out than what I had on the drawing board previously. Needs some tweaking and troubleshooting and testing but I think it'll work.
Loading is very likely broken in this commit.
Probably some runtime errors too.

I should really be doing this on a separate branch but this isn't a
release so yolo

Just wanted to get this progress saved

also renamed fastbooty function boi to reflect what it actually operates on and does

-----

Changed to GPL License.

Renamed project to CANTBoot. had this rename on my mind for a while